Division into different translation units
