+++ erohin.vladimir/S2/bracket.hpp
@@ -1,0 +1,18 @@
+#ifndef BRACKET_HPP
+#define BRACKET_HPP
+
+namespace erohin
+{
+  enum bracket_t
+  {
+    OPEN_BRACKET,
+    CLOSE_BRACKET
+  };
+
+  struct Bracket
+  {
+    bracket_t bracket;
+  };
+}
+
+#endif
