+++ erohin.vladimir/S2/main.cpp
@@ -2,6 +2,7 @@
 #include <sstream>
 #include <string>
 #include <stdexcept>
+#include "token.hpp"
 
 int main()
 {
