+++ erohin.vladimir/S2/operand.hpp
@@ -1,0 +1,12 @@
+#ifndef OPERAND_HPP
+#define OPERAND_HPP
+
+namespace erohin
+{
+  struct Operand
+  {
+    long long numeric;
+  };
+}
+
+#endif
