+++ erohin.vladimir/S2/token.hpp
@@ -1,6 +1,10 @@
 #ifndef TOKEN_HPP
 #define TOKEN_HPP
 
+#include "operand.hpp"
+#include "operator.hpp"
+#include "bracket.hpp"
+
 namespace erohin
 {
   enum token_identifier_t
@@ -10,34 +14,9 @@ namespace erohin
     BRACKET_TYPE
   };
 
-  struct Bracket
-  {
-    enum bracket_t
-    {
-      OPEN_BRACKET_TYPE,
-      CLOSE_BRACKET_TYPE
-    };
-    bracket_t bracket;
-  };
-
-  enum operator_t
-  {
-    OPERATOR_MINUS = '-',
-    OPERATOR_PLUS = '+',
-    OPERATOR_MULTIPLY = '*',
-    OPERATOR_DIVIDE = '/',
-    OPERATOR_MOD = '%'
-  };
-
-  struct Operator
-  {
-    operator_t operation;
-    unsigned short int priority;
-  };
-
   union token_t
   {
-    unsigned long long operand;
+    Operand operand;
     Operator operation;
     Bracket bracket;
   };
@@ -47,8 +26,6 @@ namespace erohin
     token_identifier_t id;
     token_t token;
   };
-
 }
-
 
 #endif
